Fixing a bug where post images were overwritten.
If you edited a blog post, and uploaded a new image, that new image
would overwrite all the asset_ids[] and unlink all the previous images.
We now keep hidden fields for all associated assets on the form so that
any new images will be ADDED to the array instead of replace it.